Move create form into form contact component
